+++ tests/regression_tests/api_categorisation_test.py
@@ -6,7 +6,7 @@ import pytest
 
 from requests import Session
 
-from sciety_labs.app.routers.api.categorisation.typing import (
+from sciety_labs.app.routers.api.classification.typing import (
     ArticleSearchResponseDict,
     CategorisationResponseDict
 )
